+++ src/Service/RelationshipService.php
@@ -36,9 +36,6 @@ class RelationshipService
 
             $relationships[$className] = [];
             foreach ($meta->associationMappings as $fieldName => $association) {
-                // @phpstan-ignore-next-line
-                $targetEntity = is_object($association) ? $association->targetEntity : $association['targetEntity'];
-
                 $relationDetails = [
                     'field' => $fieldName,
                     'type' => $association['type'],
@@ -47,9 +44,7 @@ class RelationshipService
                 if (AnalysisMode::DELETIONS === $mode) {
                     $deletions = [];
 
-                    // @phpstan-ignore-next-line
-                    $orphanRemoval = is_object($association) ? $association->orphanRemoval : $association['orphanRemoval'];
-                    if ($orphanRemoval) {
+                    if (isset($association['orphanRemoval']) && $association['orphanRemoval']) {
                         $deletions[] = [
                             'type' => DeletionType::ORPHAN_REMOVAL,
                             'level' => Level::ORM,
@@ -57,9 +52,7 @@ class RelationshipService
                         ];
                     }
 
-                    // @phpstan-ignore-next-line
-                    $cascade = is_object($association) ? $association->cascade : $association['cascade'];
-                    if ($cascade && in_array('remove', $cascade, true)) {
+                    if ($association['cascade'] && in_array('remove', $association['cascade'], true)) {
                         $deletions[] = [
                             'type' => DeletionType::CASCADE,
                             'level' => Level::ORM,
@@ -67,30 +60,28 @@ class RelationshipService
                         ];
                     }
 
-                    // @phpstan-ignore-next-line
-                    if (is_object($association) && property_exists($association, 'joinColumns')) {
-                        $joinColumns = $association->joinColumns;
-                    } elseif (is_array($association) && array_key_exists('joinColumns', $association)) { // @phpstan-ignore-line
+                    if (isset($association['joinColumns'])) {
                         $joinColumns = $association['joinColumns'];
                     } else {
                         $joinColumns = [];
                     }
                     if (!empty($joinColumns)) {
-                        // @phpstan-ignore-next-line
-                        $onDelete = is_object($joinColumns[0]) ? $joinColumns[0]->onDelete : $joinColumns[0]['onDelete'];
-                        if (!empty($onDelete)) {
-                            $deletions[] = [
-                                'type' => DeletionType::ON_DELETE,
-                                'level' => Level::DATABASE,
-                                'value' => $onDelete,
-                            ];
+                        if (isset($joinColumns[0]['onDelete'])) {
+                            $onDelete = $joinColumns[0]['onDelete'];
+                            if (!empty($onDelete)) {
+                                $deletions[] = [
+                                    'type' => DeletionType::ON_DELETE,
+                                    'level' => Level::DATABASE,
+                                    'value' => $onDelete,
+                                ];
+                            }
                         }
                     }
 
                     $relationDetails['deletions'] = $deletions;
                 }
 
-                $relationships[$className][$targetEntity] = $relationDetails;
+                $relationships[$className][$association['targetEntity']] = $relationDetails;
             }
         }