+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.1] - 2024-07-25
+
+### Changed
+- Switch from GPL-V3 to MIT Licence
+
+### Fixed
+- Fix a warning for onDelete property
 
 ## [0.5.0] - 2024-07-24